requirements.txt
chromadb-client
sentence-transformers==2.6.0
google-generativeai==0.8.3
torch==2.3.0
numpy<2.0.0
protobuf<5.0.0
pydantic==1.10.13
